Clean before refactoring
