+++ README.md
@@ -1,8 +1,12 @@
 # Sapphire UI Kit
 
-Sapphire UI Kit contains new widgets and themes for [Iced](https://iced.rs). Support accent colors. *Fork of [prettygooey](https://github.com/pieterdd/prettygooey)*. Sapphire UI Kit is primarily focused on the regular desktop (Linux, Windows, macOS).
+Sapphire UI Kit contains new themed components for [Iced](https://iced.rs). *Uses color schemes from [prettygooey](https://github.com/pieterdd/prettygooey)*. Sapphire UI Kit is primarily focused on the regular desktop (Linux, Windows, macOS).
 
 ![](assets/demo.png)
+
+## Structure
+
+The Crate is divided into two large parts: [`theme`](src/theme.rs) and [`widgets`](src/widgets.rs). The `theme` module contains only a styles, while `widgets` contains implementations of these widgets (as functions). Most functions are wrappers over Iced functions, but with modified styles applied. Sometimes it is allowed to change other non-styles parameters (such as widget sizes, fonts, etc.).
 
 ## Installation
 
@@ -101,10 +105,10 @@ impl Sandbox for Text {
 - [ ] New styles for:
     - [ ] Tabs (from `iced_aw`)
     - [ ] TabBar (from `iced_aw`)
-    - [ ] scroll panel
-    - [ ] checkboxes
+    - [X] scroll panel
+    - [X] checkboxes
     - [X] radio buttons (!!!)
-    - [ ] comboboxes
+    - [X] comboboxes
     - [ ] sliders
     - [ ] progress bars;
     - [ ] text editors;
@@ -117,11 +121,12 @@ impl Sandbox for Text {
     - [ ] File Dialog widget:
         - [ ] Select single file/dir;
         - [ ] Select multiple files/dirs;
-- [ ] Rewrite API
+- [X] Rewrite API
 
 ## Used in projects
 
 - [resistor](https://github.com/mskrasnov/resistor) -  Программа для вычисления сопротивления резисторов по цветам их маркировки;
+- [Cavaletto](https://github.com/mskrasnov/cavaletto) - AI program for image generation;
 
 ## License
