--- ROADMAP.md
@@ -1,25 +1,0 @@
-# Roadmap
-
-## 1.1.0
-
-- [ ] Additional styles for widgets:
-  - [ ] scrollbar
-  - [ ] toggler
-  - [ ] tooltip
-  - [ ] combo_box
-- [ ] Additional widgets:
-  - [ ] File browser
-
-## 1.0.0
-
-- [ ] Rewrite API;
-- [ ] «About Program» widget;
-- [ ] Additional styles for widgets from `iced_aw` crate:
-  - [ ] Tabs
-  - [ ] TabBar
-
-## 0.1.0
-
-- [ ] `checkbox` module;
-- [X] `radio` module;
-- [ ] `slider` module;
