--- src/common.rs
@@ -1,8 +1,0 @@
-use iced::Color;
-use iced::Background;
-
-pub(crate) static TEXT_COLOR_DEFAULT: Color = Color::from_rgb(0.95, 0.95, 0.95);
-pub(crate) static TEXT_COLOR_HOVER: Color = Color::WHITE;
-pub(crate) static TEXT_COLOR_PRESSED: Color = Color::from_rgb(0.7, 0.7, 0.7);
-pub(crate) static TEXT_COLOR_DISABLED: Color = Color::from_rgb(0.4, 0.4, 0.4);
-pub(crate) static FILL_DISABLED: Background = Background::Color(Color::from_rgb(0.2, 0.2, 0.2));
