+++ src/lib.rs
@@ -1,17 +1,32 @@
+//! # Sapphire UI Kit
+//!
+//! Sapphire UI Kit contains new themed components for [Iced](https://iced.rs).
+//! *Uses color schemes from [prettygooey](https://github.com/pieterdd/prettygooey)*.
+//! Sapphire UI Kit is primarily focused on the regular desktop (Linux, Windows,
+//! macOS).
+//!
+//! ## Structure
+//!
+//! The Crate is divided into two large parts: [`theme`] and [`widgets`]. The
+//! `theme` module contains only a styles, while `widgets` contains implementations
+//! of these widgets (as functions). Most functions are wrappers over Iced
+//! functions, but with modified styles applied. Sometimes it is allowed to change
+//! other non-styles parameters (such as widget sizes, fonts, etc.).
+//!
+//! Check the [examples](https://github.com/mskrasnov/sapphire_ui/tree/refact/examples)
+//! for more details!
+//!
+//! ## License
+//!
+//! Sapphire UI Kit distributed under MIT license.
+//!
+//! ## Support me
+//!
+//! Users from Russian and Belarus can use Sberbank: 2202206252335406 ...
+//!
+//! ... or use [Boosty](https://boosty.to/linux-for-arm/donate) (don't be surprised
+//! by the address - at first the page on boosty was for my Linux for ARM project,
+//! but in the future I started using it for other projects as well.).
+
 pub mod theme;
-
-#[cfg(feature = "toml-themes")]
-pub mod theme_config;
-
-pub mod accent;
-pub mod button;
-pub mod card;
-pub mod checkbox;
-pub mod header;
-pub mod radio;
-pub mod scrollable;
-pub mod text;
-pub mod text_input;
-
-pub(crate) mod common;
-pub mod primary_container;
+pub mod widgets;
